requirements.txt
Flask==2.0.1
pytube==10.9.3
ffmpeg-python==0.2.0
